move functions into actions file
